finished first implementation of facades

+++ src/main/java/DAO/CompaniesDAO.java
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 public interface CompaniesDAO {
     boolean isCompanyExists(String email, String password);
     boolean    addCompany(Company company);
-    void deleteCompany(int companyID);
+    void deleteCompany(long companyID);
     ArrayList<Company> getAllCompanies();
     void updateCompany(Company company);
-    Company getOneCompany(int companyID);
+    Company getOneCompany(long companyID);
 }
