+++ src/main/java/DAO/CustomerDAO.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
 public interface CustomerDAO {
     boolean isCustomerExists(String email, String password);
     boolean addCustomer(Customer customer);
-    void deleteCustomer(int customerID);
+    void deleteCustomer(long customerID);
     ArrayList<Customer> getAllCustomers();
-    Customer getOneCustomer(int customerID);
+    Customer getOneCustomer(long customerID);
     void updateCustomer(Customer customer);
 }
